+++ domains/report/reportService.go
@@ -3,6 +3,8 @@ package report
 import (
 	"context"
 	"errors"
+	"log"
+	"twoBinPJ/domains/user"
 )
 
 type ReportService struct {
@@ -17,15 +19,33 @@ func (r *ReportService) GetReportByNameService(name string) (*Report, error) {
 	return r.Repository.GetReportByName(name)
 }
 
-func (r *ReportService) ShowTheReportByID(ctx context.Context, id int) (*Report, error) {
+func (r *ReportService) ShowTheReportByID(id int) (*Report, error) {
 	return r.Repository.GetReportByID(id)
 }
 
 func (r *ReportService) CreateReportService(ctx context.Context, name, description, comments, seriousness string) (*Report, error) {
-	return r.Repository.CreateReport(name, description, comments, seriousness)
+	user, err := user.ForContext(ctx)
+	if err != nil {
+		log.Printf("token is incorrect or wrong: %s", err)
+		return nil, errors.New("INITIALIZING_TOKEN_ERROR")
+	}
+	if user.Role != 1 {
+		log.Printf("You do not have access to create report: %s", err)
+		return nil, errors.New("CREATE_REPORT_ERROR")
+	}
+	return r.Repository.CreateReport(name, description, comments, seriousness, user.Id)
 }
 
 func (r *ReportService) UpdateReport(name, description, comments, seriousness *string, report *Report, ctx context.Context) error {
+	user, err := user.ForContext(ctx)
+	if err != nil {
+		log.Printf("token is incorrect or wrong: %s", err)
+		return errors.New("INITIALIZING_TOKEN_ERROR")
+	}
+	if user.Role != 1 {
+		log.Printf("You do not have access to update report: %s", err)
+		return errors.New("UPDATE_REPORT_ERROR")
+	}
 	didUpdateName := false
 	didUpdateDescription := false
 	didUpdateComments := false
@@ -55,7 +75,7 @@ func (r *ReportService) UpdateReport(name, description, comments, seriousness *s
 		return errors.New("no update done")
 	}
 
-	err := r.Repository.UpdateReport(name, description, comments, seriousness, report.ID, didUpdateComments)
+	err = r.Repository.UpdateReport(name, description, comments, seriousness, report.ID, didUpdateComments)
 
 	if err != nil {
 		return err
@@ -64,5 +84,14 @@ func (r *ReportService) UpdateReport(name, description, comments, seriousness *s
 }
 
 func (r *ReportService) DeleteReport(ctx context.Context, id int) error {
+	user, err := user.ForContext(ctx)
+	if err != nil {
+		log.Printf("token is incorrect or wrong: %s", err)
+		return errors.New("INITIALIZING_TOKEN_ERROR")
+	}
+	if user.Role != 1 {
+		log.Printf("You do not have access to delete report: %s", err)
+		return errors.New("DELETE_REPORT_ERROR")
+	}
 	return r.Repository.DeleteReport(id)
 }
